Adjusted jenkins_slave setup script for Debian
Rbenv seems to require g++ to build jRuby

Script tried to bundle install in a place that did not even exit because the gem was never installed.

Gherkin doesn't like it if that gem isn't reasonably up to date.

Revert "Gherkin doesn't like it if that gem isn't reasonably up to date."

This reverts commit 119e16cfc87d2793853fd9e46880f5c1e09702f2.

Conflict resolution.

Actually install the ntp package.

Prevent duplicate sources.list entry.

Add MySQL to the slave (for leihs). We should extract this script into its own repo.

Install ntpdate, so it actually works when you try to use it.

Sorry, we don't have sudo.