Decorator: port "Polybox" UI to React
Decorator:
React Component to decorate the list in either the simple list form or as the full-blown "Box" with filter, sort, and display config. All params are handled for server-side rendering, UI can enhance when JS loads.

Views:
Use the decorator, MediaEntryIndex works again

Filter-Demo is now in Styleguide(!)

React:
-  fix RailsForm
-  dump presenters in view helper, always 'get' parameter as in haml