ci: add 'WIP' versions of test jobs
The current Cider-CI setup is optimized for the normal use case, "test for delivery".
When using the CI as a test runner while developing, this creates a lot of overhead for developers since every single commit needs to be treated as if it were ready to be merged (case in point: 'precompiling assets').

Solution: more jobs!

- 2 extra jobs: dev_feature-tests, dev_unti-tests
- they are dependend upon by anything!
- only triggered on /.*_dev_.*/ branches
- rest of the setup does not change at all, /.*_master_.*/ branches work exactly like before