chore: spec: fix autocomplete
Can't be tested at all anymore using mouse-based interaction 😢

Reason: The autocomplete list is hidden when attempting to click, because 'focus' is lost.
This is not a capybara, but a WebDriver issue – confirmed that using the native API has the same behaviour even when sending mouse movements and clicks as one batched action. It also behaves the same locally and in CI.

Solution: Don't use mouse interaction, but rather the keyboard support.

(Also cleaned up the shared helper usage)