feat: ui: permissionsForm: confirm leaving
example to be used for all other forms

regardings tests: this is hard to test because
of issues between capybara and webdriver when closing windows.
It's not clear if this is a problem i general or just with our setup.
It also behaving differently locally and in CI, which makes it even harder to debug.
For this reason, all test cases *must* be in different specs,
1 is left pending because it won't work even then.

--- squashed ---

feat: ui: lib: router: #confirmNavigation

listen to all the ways a user might leave the page.
should be provided with a `check` func which checks if confirmation is needed.