fix: ui: CollectionShow: use tab ids instead of indexes
id of every relates exactly to the name of controller action,
so just use that instead of indexes.

fixes a bug where the wrong tab was selected because
the dynamic display of the ‘relations’ tab broke the index-based approach.

also move rendering for all tabs into a partial
to make it more clear that always the same template is used.