feat: ui: lib: appRequest (xhr helper)
We want to have a xhr helper which handles as much cases as possible.

Even if it that means lots of parameters and checks, I think it's worth it to have a central helper because XHR is a tricky topic (and treating HTTP erros as "app errors" is too easily forgotten).
(In case any "really special" cases crop up I'm open to have more helpers, if they wrap the base one)

there are some special xhr helpers already, needs cleanup.

main differences with plain XHR:
- auto-authToken
- error/success according to HTTP status
- auto-parse JSON answers