fix: presenters: MediaResources: any?/empty?
the idea of those helpers was to make it easier to see if collection has any items, *regardless* of pagination settings.

must have broken some time ago and did exactly the wrong thing (asking if the *current page* has any, instead of "all pages".

fixes the "conditional relations tab" on MediaEntry and Collection, when a later page is displayed.