fix: ui: lib: setUrlParams
fix `path` vs `pathname` bug.
the idea is to use the url spec parsing format for internal consistency, but it was wrong in the backend.
thus lib had a bug which covered that up.

cleaned it all up.

fun fact: Rail's `url_for` has an option to only return the *pathname* which is called *only_path*.