spec: add presenter examples for InputMetaDatum
only keywords, needed for an overview of the possible combination of features and how the input should behave.

a ruby module defines 1 method which returns an example with title, description and full props, mainly a presenter dump.

currently those examples are only used in the styleguide.

dump file build locally with:
./spec/_support/presenter_examples/meta_datum_input_examples.rb | tail -n+2 > spec/_support/presenter_examples/meta_datum_input_examples.yaml

possible extensions:
- use same examples for MetaDatumInput acceptance test
- use same dump for UI snapshot testing