refactor: presenter: MediaResources selection
calling #any? on the AR collection did trigger a weird AREL bug (only in prod!)

inspecting the implementation of #any? it turns out that it queries "COUNT" (used to be "EXISTS"). Since the (total_)count in question is already used in this presenter, so it's more performant to memo that and use it for #any?