fix: ui: VideoPlayer default sizing
reduces "styling flash" when loading, because final sizes/aspect ration is not known until the video has loaded.

video.js dynamically adds those styles to the DOM, but we don't want to wait for it to load.

assumes an aspect ration of 16:9, which removes the flicker for common cases and makes it less jaring for others.

also makes the error message readable again