fix: InputTextDate: fix missing inputs
fixes 2 cases where the last input got lost:
- last characer of an input field
- second calendar selection of a duration

also fix: style: DayPicker center caption

---

bug details:
state handling in Input component was broken,
it needs to update on new props (when the form updates).

in all other inputs this is not a concern since `props.values` is always used.
the complexity of the double-datepicker makes it worth to handle this situation
and use internal state in the rest of the component(s).