feat: ui: persist list config in user session
layout and order are persisted to the visitors session (no need to be logged in)
whenever they are changed via a parameter from their defaults.
those settings are then always applied as long as the session is active,
with the exception if "saved settings" for Sets, which have higher priority.

the persistance happens as a side effect whenever the parameters are read server-side.
in some cases those parameters are set applied async, so the settings
are explicitly set via a AJAX PATCH